型集合の外へ広げた被演算子を断る

+++ tests/ComputeWeave.D2D1.Tests/D2D1TransformMapperTests.cs
@@ -302,7 +302,7 @@ public partial class D2D1TransformMapperTests
 
             // Calculate the distance from the center, to adjust the fade out
             float distance = Hlsl.Distance(xy, center);
-            float falloff = radius / 2u;
+            float falloff = (uint)radius / 2u;
             float alpha = 0;
 
             // If the pixel is close enough, keep it fully visible
